+++ Src/app/application.hpp
@@ -2,6 +2,8 @@
 
 #include "app/cli_options.hpp"
 #include "app/server_runtime.hpp"
+#include "ports/i_frame_log.hpp"
+#include "ports/i_historian.hpp"
 
 #include <memory>
 
@@ -20,6 +22,8 @@ private:
     std::unique_ptr<ports::ILog> log_;
     std::unique_ptr<ports::IClock> clock_;
     std::unique_ptr<ports::IMetrics> metrics_;
+    std::unique_ptr<ports::IHistorian> historian_;
+    std::unique_ptr<ports::IFrameLog> frame_log_;
     std::unique_ptr<ServerRuntime> runtime_;
 };